refactor: split Ink parser into InkDocument.js; rename KeyBuilder to KeyLoaderInk
